Add example.settings.json and update project file
The `LibreWms.FunctionApp.csproj` file has been modified to ensure that `example.settings.json` is copied to the output directory during the build process but excluded from the publish output.

A new `example.settings.json` file has been introduced, which includes settings for local development, such as `AzureWebJobsStorage` set to use development storage, the runtime for Azure Functions, and a placeholder for the `ProductsDb` connection string.